Give people working with a framework some pointers
Because I had to do this exact thing, and it was a serious pain to figure all of that out from the separate sources. It would be helpful if the buildpack would know to do this, but in the meantime having a description of the steps involved might be handy.