Update docs to add custom trusted certs
The current documentation says that the buildpack will automatically
include the certificates configured in bosh using [1] in the
director.trusted_certs property of the bosh director.

But this does not appear to be the case. We were not able to find any
logic related to implement such behaviour, and it does violate the
principle of the container, where the staging or runtime will not be
able to access the Host VM filesystem. We tested it and it does not seem
to work.

Instead, CF provides two strategies to inject custom certs into the
application container, as described in [2], by setting the CF manifest
properties by either setting cflinuxfs2-rootfs.trusted_certs or
diego.rep.trusted_certs.

This second approach does work[3][4] and makes sense to be the
recommeded way.

In this PR we update the corresponding buildpack documentation to point
to the right configuration.

[1] http://bosh.io/docs/trusted-certs.html
[2] https://docs.cloudfoundry.org/adminguide/trusted-system-certificates.html
[3] https://github.com/cloudfoundry/cflinuxfs2-release/blob/master/jobs/cflinuxfs2-rootfs-setup/templates/pre-start#L16-L41
[4] https://github.com/cloudfoundry/diego-release/blob/21a42ca5ddbfb784598abb35597cf44477f8ae27/jobs/rep/templates/rep_ctl.erb#L54-L67