Update Staticfile location instructions
The Staticfile location instructions were unclear in telling users to place the file at the root because root directory implies root of the repository.  On further investigation this really means in the _build_ directory.  See this issue for more information - it instructs users to place the Staticfile in the build directory also:  https://github.com/cloudfoundry/staticfile-buildpack/issues/148