document `package main` in a subdirectory
[#166781190](https://www.pivotaltracker.com/story/show/166781190)

We were pushing a Go app and spent a bit of time trying to figure this out.

We are not sure if this should be under the Go Modules section or some other
part of the page. We tested with Go Modules so we are putting it under that 
header. I am not sure of other packaging systems have their own way of 
specifying subdirectory main packages.

cc: @matt-royal